ASACORE-3472 PermissionConfigurator Statement Test
Fixed a typo in the JNI code.

Unit tests cover all functions of PermissionConfigurator
except getConnectedPeerPublicKey, since this function is used with
CredentialAccessor and since CredentialAccessor is not available in the
Java API. Credential Accessor extracts the bus guid and the bus guid
is used as a parameter for getConnectedPeerPublicKey.

Change-Id: Ie23d80e3864538413d63efb16104888270a556d9
Signed-off-by: George Tang <gtang@affinegy.com>