ASACORE-3502 bad_alloc error
Analysis: this error occurred because the code was trying to access
memory that had already been cleared. The vector object in
alljoyn_java/jni/PermissionConfigurator.cc is local to the
function, so the memory allocated to the vector object was unusable
outside it.

Updated the test to use the fix.

Change-Id: Ifdf11b5aa8e417aaa5633f7d8b8f7c20bf31f46d
Signed-off-by: George Tang <gtang@affinegy.com>