ASACORE-3545 Junit tests have signal ArgDefs with wrong direction
There were several cases in the junit tests where a SignalDef
definition was created that had an ArgDef defined with
direction='in'. This is wrong, because signal arguments
must be defined with direction='out'.

To correct this, the constructor ArgDef(name,type,direction)
will be used instead of the ArgDef(name,type) constructor,
so that the direction can explicitely be specified.
The ArgDef(name,type) constructor now has its comment
updated to warn against using it to create signal method args.

Change-Id: Id0b424901ca571eb519fa961bad5e038b15214b3
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>