ASACORE-3545 Junit tests have signal ArgDefs with wrong direction
There were several cases in the junit tests where a SignalDef
definition was created that had an ArgDef defined with
direction='in'. This is wrong, because signal arguments
must be defined with direction='out'.

To correct this, the constructor ArgDef(name,type) will
default the direction to "out". This is what it should
have been all along.

NOTES:

Now when creating a signal ArgDef, the ArgDef(name,type)
constructor can be used, which results in the correct
direction used for signal args by default.

When creating a method ArgDef, the ArgDef(name,type,direction)
constructor should be used. The direction should be explicitly
specified when defining method args.

Change-Id: Id0b424901ca571eb519fa961bad5e038b15214b3
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>