ASACORE-3567: Unable to build JNI SimpleClient/SimpleService in Android Studio
Fixed bug in Android.mk where ALLJOYN_DIST environment variable is not used
when set (bug in how "ifndef" was used such that it was always true).

Fixed bug in Java code where Client would display duplicate entries
for a  found simple service.

Added "how to" notes to the JIRA ticket, describing how to build
JNI samples in Android Studio as well as from the Linux command-line.

The difficulty with bilding in Android Studio is that the existing
Android JNI Samples exist in the repo as Eclipse projects, which can't
simply be imported into Android Studio and immediately work; i.e.
after importing into Android Studio there are numerious manual steps
in fixing the configuation before the JNI samples can be built
successfully. However, submiting Android Studio versions of these
JNI samples to the repo seems to be (?) beyond the 16.10a scope.

Change-Id: I00035ce7612b8d1e99f2ab6a127f99d61d5f76d3
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>