ASACORE-3619: Segfault when BusAttachment has concurrency set to 0
This fix is based on ASACORE-3512: Remove BusAttachment's
concurrency limit. ASACORE-3512 was changing API but this change
is just changing behaviour when BusAttachment is created with
concurrency set to 0. Until now it was causing segfault in Timer.cc
but with this changes when concurrency is set to 0, there is no
concurrency limit and handlers are processed as it was
implemented in ASACORE-3512.

Change-Id: I57696700d905f4b364af3e6da8e41f8bd6e1b30a
Signed-off-by: Krystian Zlomek <kzlomek@affinegy.com>