ASACORE-2716: Adding AuthMechanism callback in AuthListener
The original function described in this issue is not suitable for the
desired test. Adding a new callback function for such purpose.

Note: The entire fix would require commits from alljoyn_core and
core/test/ repos, this is only the alljoyn_core side of change and must
go in first before core/test changes could pass the CI tests.

Change-Id: I24a357f3d651fef52c1fabd3bbfb1c387a397396
Signed-off-by: Vincent Du <vdu@affinegy.com>