ASACORE-3537: PropertyChangedEmitter not usable with DynamicBusObject (Java)
Updated PropertryChangedEmitter so that it can be used to emit
property change signals for dynamic bus objects.

Fixed several problems in related classes that were affecting
the working of the updates in PropertyChagnedEmitter...

Fixed problem in AbstractDynamicBusObject. Its member variables
are all final, but the class constructor only asserts that some
of the constructor inputs are non-null. Ex: It was possible to
send null bus value to constructor, meaning that its mBus member
could never be initialized (since the constructor is the
only way to initialize this class' final members).

Fixed problem in DynamicBusObject interface. It didn't define
getBus() method, but realistically a dynamic bus object instance
would need to know its bus attachement if it is going to do
anything interesting (e.g. dynamic bus objects can't know which
method/signal/property is being called unless it can
access the bus attachment's MessageContext).

Fixed problem in SignalEmitter where adding the Properties
interface to the list of supported interfaces will fail if
the list being updated was defined by the caller as a
fixed-sized list.

Change-Id: I097f10c10757cbc0560869c738d6a641dbbf1bb7
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>